+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * Add a limit to email size.
 ### Fix
 * Normalize email in find
+* Ensure that email is compliant with Regex
 * Check filter for phone listing
 * Don't let admin remove a user
 * Ensure that car is linked to the current selected campus